# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type CopyMessageType = typeof CopyMessageType[keyof typeof CopyMessageType];


export const CopyMessageType = {
  venda: 'venda',
  followup: 'followup',
  urgencia: 'urgencia',
  posVenda: 'posVenda',
  objecao: 'objecao',
} as const;

export type CopyTone = typeof CopyTone[keyof typeof CopyTone];


export const CopyTone = {
  profissional: 'profissional',
  amigavel: 'amigavel',
  direto: 'direto',
  emocional: 'emocional',
} as const;

export interface Copy {
  id: number;
  clientName: string;
  product: string;
  /** @nullable */
  value?: string | null;
  /** @nullable */
  context?: string | null;
  messageType: CopyMessageType;
  tone?: CopyTone;
  generatedText: string;
  createdAt: string;
}

export type CopyInputMessageType = typeof CopyInputMessageType[keyof typeof CopyInputMessageType];


export const CopyInputMessageType = {
  venda: 'venda',
  followup: 'followup',
  urgencia: 'urgencia',
  posVenda: 'posVenda',
  objecao: 'objecao',
} as const;

export type CopyInputTone = typeof CopyInputTone[keyof typeof CopyInputTone];


export const CopyInputTone = {
  profissional: 'profissional',
  amigavel: 'amigavel',
  direto: 'direto',
  emocional: 'emocional',
} as const;

export interface CopyInput {
  /** @minLength 1 */
  clientName: string;
  /** @minLength 1 */
  product: string;
  value?: string;
  context?: string;
  messageType: CopyInputMessageType;
  tone?: CopyInputTone;
}

export type CopyStatsByType = {[key: string]: number};

export interface CopyStats {
  total: number;
  byType: CopyStatsByType;
  todayCount: number;
}

export interface PaginatedCopies {
  copies: Copy[];
  total: number;
  page: number;
  limit: number;
}

export interface User {
  id: number;
  name: string;
  email: string;
  plan: string;
}

export interface AuthResponse {
  user: User;
  token: string;
}

export interface AIResponse {
  id: number;
  generatedText: string;
  createdAt: string;
}

export interface ImprovedCopyResponse {
  id: number;
  originalText: string;
  improvedText: string;
  createdAt: string;
}

export interface FollowUpResponse {
  id: number;
  generatedText: string;
  followUpStrategy: string;
  createdAt: string;
}