# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: CopyZap AI - WhatsApp Sales Copy Generator API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication operations
  - name: copies
    description: Copy generation operations
  - name: ai
    description: Advanced AI features
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register new user
      description: Create a new user account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "409":
          description: Email already exists

  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      description: Authenticate user and get token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials

  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      description: Returns the authenticated user's data
      security:
        - bearerAuth: []
      responses:
        "200":
          description: User data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  /copies:
    get:
      operationId: listCopies
      tags: [copies]
      summary: List generated copies
      description: Returns paginated list of copies for the authenticated user
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 50
      responses:
        "200":
          description: Paginated list of copies
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedCopies"
    post:
      operationId: generateCopy
      tags: [copies]
      summary: Generate a copy with AI
      description: Uses AI to generate a WhatsApp sales message
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CopyInput"
      responses:
        "201":
          description: Generated copy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Copy"
        "429":
          description: Daily limit reached

  /copies/{id}:
    delete:
      operationId: deleteCopy
      tags: [copies]
      summary: Delete a copy
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /copies/stats:
    get:
      operationId: getCopyStats
      tags: [copies]
      summary: Get copy statistics
      description: Returns counts per message type and recent activity for the user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CopyStats"

  /ai/seller-mode:
    post:
      operationId: sellerMode
      tags: [ai]
      summary: AI Seller Mode
      description: Describe a client scenario, AI responds as a human seller
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SellerModeInput"
      responses:
        "201":
          description: Generated seller response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AIResponse"

  /ai/improve-copy:
    post:
      operationId: improveCopy
      tags: [ai]
      summary: Improve a copy
      description: Paste a message and AI makes it more persuasive
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ImproveCopyInput"
      responses:
        "201":
          description: Improved copy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImprovedCopyResponse"

  /ai/follow-up:
    post:
      operationId: generateFollowUp
      tags: [ai]
      summary: Generate follow-up message
      description: Generate a follow-up message for a lead that went cold
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FollowUpInput"
      responses:
        "201":
          description: Generated follow-up
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FollowUpResponse"

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    RegisterInput:
      type: object
      required:
        - name
        - email
        - password
      properties:
        name:
          type: string
          minLength: 2
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6

    LoginInput:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
        password:
          type: string

    AuthResponse:
      type: object
      required:
        - user
        - token
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string

    User:
      type: object
      required:
        - id
        - name
        - email
        - plan
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        plan:
          type: string
          enum: [free, pro]

    Copy:
      type: object
      required:
        - id
        - userId
        - clientName
        - product
        - messageType
        - generatedText
        - createdAt
      properties:
        id:
          type: integer
        userId:
          type: integer
        clientName:
          type: string
        product:
          type: string
        value:
          type: ["string", "null"]
        context:
          type: ["string", "null"]
        messageType:
          type: string
          enum: [venda, followup, urgencia, posVenda, objecao]
        tone:
          type: string
          enum: [profissional, amigavel, direto, emocional]
        generatedText:
          type: string
        createdAt:
          type: string

    CopyInput:
      type: object
      required:
        - clientName
        - product
        - messageType
      properties:
        clientName:
          type: string
          minLength: 1
        product:
          type: string
          minLength: 1
        value:
          type: string
        context:
          type: string
        messageType:
          type: string
          enum: [venda, followup, urgencia, posVenda, objecao]
        tone:
          type: string
          enum: [profissional, amigavel, direto, emocional]

    PaginatedCopies:
      type: object
      required:
        - copies
        - total
        - page
        - limit
      properties:
        copies:
          type: array
          items:
            $ref: "#/components/schemas/Copy"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer

    CopyStats:
      type: object
      required:
        - total
        - byType
        - todayCount
      properties:
        total:
          type: integer
        byType:
          type: object
          additionalProperties:
            type: integer
        todayCount:
          type: integer

    SellerModeInput:
      type: object
      required:
        - clientDescription
        - product
      properties:
        clientDescription:
          type: string
          minLength: 10
        product:
          type: string
          minLength: 1
        context:
          type: string

    AIResponse:
      type: object
      required:
        - id
        - generatedText
        - createdAt
      properties:
        id:
          type: integer
        generatedText:
          type: string
        createdAt:
          type: string

    ImproveCopyInput:
      type: object
      required:
        - originalMessage
      properties:
        originalMessage:
          type: string
          minLength: 10
        goal:
          type: string
          enum: [venda, profissional, amigavel, urgencia, emocional]
        additionalInstructions:
          type: string

    ImprovedCopyResponse:
      type: object
      required:
        - id
        - originalText
        - improvedText
        - createdAt
      properties:
        id:
          type: integer
        originalText:
          type: string
        improvedText:
          type: string
        createdAt:
          type: string

    FollowUpInput:
      type: object
      required:
        - clientName
        - product
        - lastInteraction
        - daysSinceLastContact
      properties:
        clientName:
          type: string
        product:
          type: string
        lastInteraction:
          type: string
          minLength: 5
        daysSinceLastContact:
          type: integer
          minimum: 0
        previousOutcome:
          type: string
          enum: [interested, no_response, objection, ghosted]

    FollowUpResponse:
      type: object
      required:
        - id
        - generatedText
        - followUpStrategy
        - createdAt
      properties:
        id:
          type: integer
        generatedText:
          type: string
        followUpStrategy:
          type: string
        createdAt:
          type: string